+++ forge/db/models/TeamBrokerClient.js
@@ -1,6 +1,6 @@
 const { DataTypes } = require('sequelize')
 
-const { hash, buildPaginationSearchClause } = require('../../../db/utils')
+const { hash, buildPaginationSearchClause } = require('../utils')
 
 module.exports = {
     name: 'TeamBrokerClient',
@@ -19,6 +19,26 @@ module.exports = {
     ],
     associations: function (M) {
         this.belongsTo(M.Team)
+    },
+    hooks: function (M, app) {
+        return {
+            beforeCreate: async (project, opts) => {
+                // if the product is licensed, we permit overage
+                const isLicensed = app.license.active()
+                if (isLicensed !== true) {
+                    const { mqttClients } = await app.license.usage('mqttClients')
+                    if (mqttClients.count >= mqttClients.limit) {
+                        throw new Error('license limit reached')
+                    }
+                }
+            },
+            afterCreate: async (project, opts) => {
+                const { mqttClients } = await app.license.usage('mqttClients')
+                if (mqttClients.count > mqttClients.limit) {
+                    await app.auditLog.Platform.platform.license.overage('system', null, mqttClients)
+                }
+            }
+        }
     },
     finders: function (M) {
         return {
