+++ forge/ee/db/models/index.js
@@ -13,8 +13,7 @@ const modelTypes = [
     'PipelineStageDevice',
     'PipelineStageDeviceGroup',
     'FlowTemplate',
-    'MFAToken',
-    'TeamBrokerClient'
+    'MFAToken'
 ]
 
 async function init (app) {
@@ -48,7 +47,7 @@ async function init (app) {
         }
         if (m.hooks) {
             if (typeof m.hooks === 'function') {
-                opts.hooks = m.hooks.call(null, app.db.models)
+                opts.hooks = m.hooks.call(null, app.db.models, app)
             } else {
                 opts.hooks = m.hooks
             }
